double int issue fixed

+++ lib/service/book_confirmation_service.dart
@@ -6,10 +6,10 @@ import 'package:qixer/service/booking_services/personalization_service.dart';
 class BookConfirmationService with ChangeNotifier {
   bool isPanelOpened = false;
 
-  double totalPriceAfterAllcalculation = 0;
-  var subTotalAfterAllCalculation = 0;
-  double totalPriceOnlineServiceAfterAllCalculation = 0;
-  var subTotalOnlineServiceAfterAllCalculation = 0;
+  double totalPriceAfterAllcalculation = 0.0;
+  var subTotalAfterAllCalculation = 0.0;
+  double totalPriceOnlineServiceAfterAllCalculation = 0.0;
+  var subTotalOnlineServiceAfterAllCalculation = 0.0;
 
   var taxPrice;
   var taxPriceOnline;
@@ -25,27 +25,26 @@ class BookConfirmationService with ChangeNotifier {
   }
 
   includedTotalPrice(List includedList) {
-    var total = 0;
+    var total = 0.0;
     for (int i = 0; i < includedList.length; i++) {
-      total =
-          total + (includedList[i]['price'] * includedList[i]['qty']) as int;
+      total = total + (includedList[i]['price'] * includedList[i]['qty']);
     }
     return total;
   }
 
   extrasTotalPrice(List extrasList) {
-    var total = 0;
+    var total = 0.0;
     for (int i = 0; i < extrasList.length; i++) {
       if (extrasList[i]['selected'] == true) {
-        total = total + (extrasList[i]['price'] * extrasList[i]['qty']) as int;
+        total = total + (extrasList[i]['price'] * extrasList[i]['qty']);
       }
     }
     return total;
   }
 
   calculateSubtotal(List includedList, List extrasList) {
-    var includedTotal = 0;
-    var extraTotal = 0;
+    var includedTotal = 0.0;
+    var extraTotal = 0.0;
     includedTotal = includedTotalPrice(includedList);
     extraTotal = extrasTotalPrice(extrasList);
     subTotalAfterAllCalculation = includedTotal + extraTotal;
@@ -54,7 +53,7 @@ class BookConfirmationService with ChangeNotifier {
   }
 
   calculateSubtotalForOnline(List extrasList) {
-    var extraTotal = 0;
+    var extraTotal = 0.0;
 
     extraTotal = extrasTotalPrice(extrasList);
     subTotalOnlineServiceAfterAllCalculation = extraTotal;
